Migrated code from Repository.java and deleted it
